273.ProductVItem-protecting relations

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -99,6 +100,16 @@ class ProductVariantController extends Controller
     {
         $variant = ProductVariant::findOrFail($id);
 
+        $variantItemCheck = ProductVariantItem::where('product_variant_id', $variant->id)
+            ->count();
+
+        if ($variantItemCheck > 0) {
+            return response([
+                'status' => 'error',
+                'message' => 'This Variant Contains item in it. Delete the variant items first for delete this variant!'
+            ]);
+        }
+        
         $variant->delete();
 
         return response([
